// providers/plugin.d.ts
// Generated by dts-bundle-generator v6.11.0

import React from 'react'

export declare type PluginApiVersion = 1
export declare type PluginInfo<T = string> = {
  apiVersion: PluginApiVersion
  type: T
  name: string
  renderIcon?: () => React.ReactNode
}
export declare type PluginConfig<T = Record<string, unknown>> = Record<
  string,
  unknown
> &
  T
export interface ISearchPlugin<T = string> {
  readonly info: PluginInfo<T>
  search(q: string, sort: SearchSort): Promise<SearchResult<T>[]>
}
export declare type SearchPluginConstructor<T = string> = new (
  config: PluginConfig
) => ISearchPlugin<T>
export declare type SearchSort = 'relevance' | 'date' | null
export declare type SearchResult<T = string> = {
  type: T
  url: string
  title?: string
  imageUrl?: string | null
  description?: string | null
  author?: {
    name: string
    url?: string | null
    imageUrl?: string | null
  }
  timestamp?: {
    createdAt: number
    updatedAt?: number | null
  }
}

export {}